Add AI-generated code changes
