+++ calculator.py
@@ -1,0 +1,11 @@
+def add(x, y):
+    """Add two numbers together.
+
+    Args:
+        x (int or float): The first number to be added.
+        y (int or float): The second number to be added.
+
+    Returns:
+        int or float: The sum of the two numbers.
+    """
+    return x + y
