+++ test_calculator.py
@@ -1,19 +1,3 @@
-FILENAME: calculator
-
-FUNCTION:
-def add(x, y):
-    """Add two numbers together.
-
-    Args:
-        x (int or float): The first number to be added.
-        y (int or float): The second number to be added.
-
-    Returns:
-        int or float: The sum of the two numbers.
-    """
-    return x + y
-
-TESTS:
 from calculator import add
 import unittest
 
@@ -27,3 +11,4 @@ class TestCalculator(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
+```
